Feedback to PR, moved analytics and library/icon to its own file and imported icons as module instead of window scope. Fixed consonant associated spacing tokens, Moved default icon images.
