+++ libs/blocks/columns/columns.css
@@ -22,7 +22,7 @@
 }
 
 .columns.l-spacing {
-  gap: var(--spacing-xl);
+  gap: var(--spacing-l);
 }
 
 .columns.s-spacing {
