+++ libs/blocks/section-metadata/section-metadata.css
@@ -26,7 +26,7 @@
 }
 
 .section.l-spacing {
-  padding: var(--spacing-xl) 0;
+  padding: var(--spacing-l) 0;
 }
 
 .section.m-spacing {
